backend/requirements.txt
paddleocr==2.9.1
paddlepaddle==2.6.2

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

pdf2image==1.16.3
pillow==10.0.0
openpyxl==3.1.5

numpy==1.26.4
scikit-image==0.22.0

pydantic>=2.7.0,<3.0.0
lxml
